feat: implemented http verbs on intake under structure controller
