+++ src/routes/adminRoutes.js
@@ -1,27 +1,14 @@
 /* eslint-disable linebreak-style */
 const express = require('express');
 
-const adminsController = require('../controller/adminsController');
+const router = express.Router();
 
-function adminRouters() {
-  const adminRouter = express.Router();
-  const controller = adminsController;
+const adminController = require('../controller/adminController');
 
-  adminRouter.route('/')
-    // .post(controller.post)
-    .get(controller.get);
+router.get('/', adminController.getAll);
+router.get('/:adminId', adminController.getSingle);
+router.post('/add', adminController.adds);
+router.patch('/:adminId', adminController.updateOne);
+router.delete('/:adminId', adminController.deletes);
 
-  adminRouter.route('/add')
-    .post(controller.post);
-
-  adminRouter.route('/:adminId')
-    .all(controller.alls)
-    .get((req, res) => {
-      res.send(req.admin);
-    })
-    .patch(controller.patch)
-    .delete(controller.deletes);
-
-  return adminRouter;
-}
-module.exports = adminRouters;
+module.exports = router;
